+++ examples/personal-assistant/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aiet/example-personal-assistant
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @aiet/adapter-openai-agents@1.0.2
+  - @aiet/core@1.0.2
+  - @aiet/domain@1.0.2
+  - @aiet/schema@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes